requirements.txt
Flask==3.1.1
matplotlib==3.10.5
pandas==2.3.1
Pillow==11.3.0
scikit-image
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
flask-cors==6.0.1
opencv-python==4.8.1.78
numpy==1.26.4

